chore(explore): Move explore spans content into spans folder
Doing some light clean up, starting by moving the spans related files into the
spans folder.
