+++ static/app/views/explore/spans/content.tsx
